feat(slurm_ops): clean up public exports
We now intend that you directly call the manager that you plan on using rather
than compose it yourself within the specific Slurm charm. This commit
makes the methods and classes that we want private have a underscore before
their name so that it is more obvious that it isn't intended to be a publicly
used object outside of the charm library.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>