feat(slurm_ops): add initial jwt key support
Add initial support for managing jwt keys with the slurm_ops charm library.
Currently relies on the cryptography library for generating an RSA key that is
used to sign Slurm communications, but we'll eventually be made to rely on a dedicated
`jwtctl` tool that can handle the key for us.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>