fix(slurm_ops): use shell variable substitution in /etc/default/sackd
Variable substitution in the environment file prevents us from
needing to set the service type to `Type=forking` in systemd.
`sackd` is started the way it should be with the `--systemd` rather
than run as a child process in the shell.

`slurmd` needs some further work around `juju-systemd-notices` before it
can use this approach as well.

Other changes:

* Clean up docstrings to be accurate and reflect the current implementation of the library.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>